backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

sqlalchemy==2.0.23
psycopg2-binary==2.9.9

pydantic==2.5.0
pydantic-settings==2.1.0
email-validator
python-dotenv==1.0.0

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

python-multipart==0.0.6

pillow>=10.4.0
opencv-python==4.8.1.78

numpy==1.24.3
scikit-learn==1.3.2

torch==2.1.1
torchvision==0.16.1

shap==0.43.0

requests==2.31.0
cloudinary==1.36.0

prometheus-client==0.19.0
mlflow==2.10.0

psutil==5.9.6

typing-extensions>=4.8.0